Implement getStyles method
This method will get all style entries associated with the given
productId.

It will store those as a result array on an object that trakcs the
product id.

It will then use the style ids it finds to find all photo entries and
sku entries associated with each of the styles found in the original db
call.

The results of that second round of queries will be appended into their
respective styles before return the compiled resultObj to the controller