Add redis caching and optimize queries
Removes sku model as it was not being utilized

All models:
Now utilize redis for caching results for increased effciency and
support for abnormal trafic
Each model now has a method that utilize the spcific endpoint details to
check the redis cache to see if there is a stored result for the current
query
All models databsee retrival methods will now cache the results of the
query so they can be easily retrieved in future queries.
Any queries that are not part of the landing page of the front end app
will only remain in the cache for a maximum of 5 seconds without being
accessed.  This will allow the server to handle extremely high volumes
to a specific endpoint should it occur.

productsModel:
getProducts method - the query has been refactored to no longer utilize
the OFFSET clause - the offset clause ignored the effciencies of indexed
records - query will now use the calculated offset to specify which
product Id to start with